refactor: wildcard exports + true polymorphism
