+++ frontend/src/assets/Logo.tsx
@@ -1,59 +1,55 @@
 import { DetailedHTMLProps, SVGAttributes } from 'react';
 
-function Logo(props: DetailedHTMLProps<SVGAttributes<SVGSVGElement>, SVGSVGElement>) {
-  return (
-    <svg {...props} version="1.1" viewBox="0 0 2650 700" xmlns="http://www.w3.org/2000/svg">
-      <path
-        fill="#00BAFF"
-        d=" M28.167587,354.868042   C35.868050,322.277985 49.972065,292.650879 69.805077,265.758789   C82.927628,247.965607 98.131035,232.223145 115.652962,218.562561   C131.263702,206.392014 147.873367,196.072845 165.742294,187.717606   C184.432465,178.978363 203.937958,172.768509 224.240341,168.938583   C238.493698,166.249786 252.793274,164.587326 267.710114,165.536499   C268.075043,170.383698 267.933899,174.714981 268.019684,179.041748   C268.273804,191.860657 264.453552,203.602448 258.894226,214.946747   C258.135773,216.494446 257.880524,218.428650 257.876312,220.185059   C257.814423,246.004288 257.708099,271.824829 257.933441,297.642242   C257.983398,303.366547 258.859131,309.169708 260.074432,314.780670   C263.005707,328.314270 269.041656,341.032593 270.181396,355.093872   C270.986725,365.029114 271.886536,374.602692 268.133484,384.404755   C266.274048,389.261200 263.411102,393.162720 261.031525,397.524963   C255.011490,408.560883 248.986923,419.616943 243.628555,430.978302   C241.860840,434.726440 240.883194,439.684753 241.686218,443.656708   C244.532364,457.734375 253.752426,466.738342 266.264832,473.438049   C276.949280,479.158997 288.487488,482.043427 299.989655,483.639587   C315.513580,485.793793 331.379211,485.832886 347.107422,486.024170   C355.194855,486.122559 363.261627,483.611389 368.446320,477.006653   C374.744568,468.983368 380.284119,460.313202 385.602966,451.588226   C390.050781,444.292114 395.935120,438.959991 404.177917,437.266785   C418.220306,434.382324 432.219757,435.305145 445.727051,440.356445   C455.229767,443.910187 464.810333,446.572266 475.095184,446.076324   C476.955322,445.986633 478.855133,446.934631 480.751221,447.334534   C486.628845,448.574188 492.485016,450.578461 498.402924,450.789642   C505.184998,451.031708 512.024231,449.671265 518.838684,449.006989   C517.117126,458.574463 515.878845,468.260193 513.572021,477.684418   C508.978210,496.452148 502.054230,514.397644 493.125244,531.572571   C484.997528,547.206177 475.573669,561.965027 464.101807,575.349854   C458.206146,582.228638 452.374390,589.230713 445.861664,595.499573   C423.955292,616.585693 399.124878,633.301758 371.161560,645.537109   C352.944733,653.507935 334.078583,659.057739 314.699463,662.859497   C297.475037,666.238586 279.938934,666.220947 262.419495,666.052856   C241.028198,665.847534 220.311996,661.983154 199.817825,656.092712   C172.245255,648.167725 146.857834,635.667358 123.902580,618.753784   C110.840797,609.129700 99.328812,597.402283 87.155746,585.961548   C88.847534,579.962341 90.492218,574.620483 92.163025,569.286804   C92.740837,567.442200 93.498398,565.647400 93.977486,563.779480   C96.512589,553.895264 98.231499,543.721008 101.700089,534.180542   C104.903587,525.369263 101.033112,513.152283 92.990562,507.287781   C87.158600,503.035187 79.954857,497.365906 82.046188,487.770020   C82.629921,485.091614 83.486923,482.393433 84.713257,479.952881   C86.567871,476.262024 89.483871,473.011475 90.840782,469.181274   C94.678528,458.348389 97.319405,447.062317 101.582314,436.415405   C110.231949,414.812439 96.112907,391.530853 81.856262,383.153473   C64.435081,372.916565 46.104771,364.226807 28.167587,354.868042  M231.298508,607.789551   C227.994736,610.190979 224.690979,612.592407 220.566452,615.590393   C221.354904,617.586487 222.064163,620.289795 223.432648,622.603088   C224.940567,625.152100 227.074356,627.327148 228.897736,629.694214   C230.708786,632.045227 232.617111,633.427368 235.891266,634.295715   C242.986343,636.177612 250.636139,637.540588 255.544830,644.366455   C257.080688,646.502197 260.127258,648.607910 262.636200,648.809265   C275.064758,649.806519 287.159760,647.564087 298.989716,643.728210   C305.123199,641.739380 311.248779,638.660461 317.493896,638.341003   C326.039612,637.903870 334.111908,636.031921 341.613861,632.671326   C346.017456,630.698730 351.531525,626.555725 352.475586,622.452515   C354.609894,613.176086 353.081573,603.602905 345.831482,596.129883   C341.771301,591.944702 337.747742,587.698669 333.385284,583.844910   C331.718292,582.372314 328.811249,580.707092 327.052490,581.202087   C317.881042,583.783142 308.378662,586.011597 299.970886,590.288330   C287.605438,596.578064 276.073639,600.619629 262.413971,593.629211   C254.923157,589.795898 247.061752,590.930542 241.235306,598.688599   C238.724380,602.032043 235.023148,604.481445 231.298508,607.789551 z"
-      />
-      <path
-        fill="#00B501"
-        d=" M231.580597,607.563232   C235.023148,604.481445 238.724380,602.032043 241.235306,598.688599   C247.061752,590.930542 254.923157,589.795898 262.413971,593.629211   C276.073639,600.619629 287.605438,596.578064 299.970886,590.288330   C308.378662,586.011597 317.881042,583.783142 327.052490,581.202087   C328.811249,580.707092 331.718292,582.372314 333.385284,583.844910   C337.747742,587.698669 341.771301,591.944702 345.831482,596.129883   C353.081573,603.602905 354.609894,613.176086 352.475586,622.452515   C351.531525,626.555725 346.017456,630.698730 341.613861,632.671326   C334.111908,636.031921 326.039612,637.903870 317.493896,638.341003   C311.248779,638.660461 305.123199,641.739380 298.989716,643.728210   C287.159760,647.564087 275.064758,649.806519 262.636200,648.809265   C260.127258,648.607910 257.080688,646.502197 255.544830,644.366455   C250.636139,637.540588 242.986343,636.177612 235.891266,634.295715   C232.617111,633.427368 230.708786,632.045227 228.897736,629.694214   C227.074356,627.327148 224.940567,625.152100 223.432648,622.603088   C222.064163,620.289795 221.354904,617.586487 220.566452,615.590393   C224.690979,612.592407 227.994736,610.190979 231.580597,607.563232  z"
-      />
-      <path
-        fill="#00BAFB"
-        d=" M519.370117,446.942352   C518.667053,446.479004 518.004944,445.764435 518.050110,445.097687   C518.615417,436.739807 519.285889,428.389099 519.921814,420.035980   C520.035339,418.545013 520.114868,417.051453 520.525513,415.319275   C520.473877,425.624054 520.106873,436.168701 519.370117,446.942352  z"
-      />
-      <path
-        fill="#666666"
-        d=" M265.009949,128.533661   C263.641663,133.968658 259.973389,131.882767 256.944641,131.403793   C252.602493,130.717087 248.323746,129.553268 243.965057,129.063889   C236.171356,128.188858 228.920319,126.208847 224.187042,119.453926   C222.951370,117.690460 222.927948,114.971542 222.672226,112.655617   C222.385757,110.061180 222.758636,107.376442 222.322601,104.820038   C221.630081,100.759674 220.155182,98.134880 214.961990,100.335991   C210.931198,102.044434 205.621414,98.917557 203.173035,94.262817   C200.985229,90.103500 201.370758,86.327354 205.277313,83.197380   C210.683334,78.866005 216.012924,74.439224 221.725922,70.021103   C232.301041,76.411461 240.159500,85.230125 247.311707,94.789009   C254.948608,104.995659 260.890991,116.084435 265.009949,128.533661  z"
-      />
-      <path
-        fill="#00B500"
-        d=" M518.900635,448.705261   C512.024231,449.671265 505.184998,451.031708 498.402924,450.789642   C492.485016,450.578461 486.628845,448.574188 480.751221,447.334534   C478.855133,446.934631 476.955322,445.986633 475.095184,446.076324   C464.810333,446.572266 455.229767,443.910187 445.727051,440.356445   C432.219757,435.305145 418.220306,434.382324 404.177917,437.266785   C395.935120,438.959991 390.050781,444.292114 385.602966,451.588226   C380.284119,460.313202 374.744568,468.983368 368.446320,477.006653   C363.261627,483.611389 355.194855,486.122559 347.107422,486.024170   C331.379211,485.832886 315.513580,485.793793 299.989655,483.639587   C288.487488,482.043427 276.949280,479.158997 266.264832,473.438049   C253.752426,466.738342 244.532364,457.734375 241.686218,443.656708   C240.883194,439.684753 241.860840,434.726440 243.628555,430.978302   C248.986923,419.616943 255.011490,408.560883 261.031525,397.524963   C263.411102,393.162720 266.274048,389.261200 268.133484,384.404755   C271.886536,374.602692 270.986725,365.029114 270.181396,355.093872   C269.041656,341.032593 263.005707,328.314270 260.074432,314.780670   C258.859131,309.169708 257.983398,303.366547 257.933441,297.642242   C257.708099,271.824829 257.814423,246.004288 257.876312,220.185059   C257.880524,218.428650 258.135773,216.494446 258.894226,214.946747   C264.453552,203.602448 268.273804,191.860657 268.019684,179.041748   C267.933899,174.714981 268.075043,170.383698 268.191406,165.603180   C293.549225,163.907318 318.108063,167.841995 342.325745,174.915939   C371.543732,183.450500 398.177368,197.115952 422.409607,215.296295   C439.351318,228.006866 454.432098,242.741333 467.451202,259.639740   C479.057983,274.704895 489.272064,290.538666 497.279541,307.743835   C507.291199,329.255249 514.930969,351.598694 517.908325,375.224213   C519.533386,388.119415 519.982788,401.162750 520.586182,414.849182   C520.114868,417.051453 520.035339,418.545013 519.921814,420.035980   C519.285889,428.389099 518.615417,436.739807 518.050110,445.097687   C518.004944,445.764435 518.667053,446.479004 519.000427,447.171356   C518.987793,447.582062 518.975159,447.992798 518.900635,448.705261  z"
-      />
-      <path
-        fill="#00B500"
-        d=" M28.086550,355.266937   C46.104771,364.226807 64.435081,372.916565 81.856262,383.153473   C96.112907,391.530853 110.231949,414.812439 101.582314,436.415405   C97.319405,447.062317 94.678528,458.348389 90.840782,469.181274   C89.483871,473.011475 86.567871,476.262024 84.713257,479.952881   C83.486923,482.393433 82.629921,485.091614 82.046188,487.770020   C79.954857,497.365906 87.158600,503.035187 92.990562,507.287781   C101.033112,513.152283 104.903587,525.369263 101.700089,534.180542   C98.231499,543.721008 96.512589,553.895264 93.977486,563.779480   C93.498398,565.647400 92.740837,567.442200 92.163025,569.286804   C90.492218,574.620483 88.847534,579.962341 87.013214,585.651123   C77.587128,573.867859 67.571922,562.236206 59.256271,549.495239   C42.765011,524.227905 32.403427,496.351013 25.836536,466.868835   C20.668997,443.669098 20.105732,420.240967 21.298513,396.829681   C22.001575,383.030304 25.670464,369.382080 28.086550,355.266937  z"
-      />
-      <path
-        fill="#7E7E7E"
-        d=" M141.069504,45.157181   C170.998398,43.143059 197.655746,52.043030 221.583542,69.703674   C216.012924,74.439224 210.683334,78.866005 205.277313,83.197380   C201.370758,86.327354 200.985229,90.103500 203.173035,94.262817   C205.621414,98.917557 210.931198,102.044434 214.961990,100.335991   C220.155182,98.134880 221.630081,100.759674 222.322601,104.820038   C222.758636,107.376442 222.385757,110.061180 222.672226,112.655617   C222.927948,114.971542 222.951370,117.690460 224.187042,119.453926   C228.920319,126.208847 236.171356,128.188858 243.965057,129.063889   C248.323746,129.553268 252.602493,130.717087 256.944641,131.403793   C259.973389,131.882767 263.641663,133.968658 265.020172,128.913849   C269.571960,138.627701 270.715302,149.070221 272.108276,159.857117   C230.363998,158.656036 191.489380,168.150406 154.878464,186.681641   C117.979553,205.358643 87.864510,232.118225 63.655178,265.882141   C48.646038,252.609314 37.576401,236.751312 29.994730,218.412704   C24.765434,205.764023 21.450434,192.673157 21.561848,178.473450   C28.906944,178.014832 35.730309,178.000168 42.553642,178.008820   C49.421242,178.017532 54.591679,175.338821 58.042816,169.137390   C60.536346,164.656723 63.285271,160.224899 66.536179,156.278671   C73.032448,148.392944 81.726234,144.451630 91.801537,142.893692   C96.917999,142.102524 101.946121,140.391693 106.837921,138.612122   C110.403694,137.314957 113.529907,134.774292 117.114983,133.571854   C128.815033,129.647644 138.108780,123.314041 144.755234,112.350838   C148.581451,106.039536 149.427597,103.695267 142.236649,97.576584   C139.087112,94.896683 135.803009,92.375427 132.597900,89.760208   C128.325790,86.274353 124.613472,75.030449 127.260376,71.859322   C130.071991,68.490868 133.400223,65.141113 137.208694,63.100632   C145.905457,58.441132 147.396515,52.827572 141.069504,45.157181  z"
-      />
-      <path
-        fill="#666666"
-        d=" M140.610550,45.097717   C147.396515,52.827572 145.905457,58.441132 137.208694,63.100632   C133.400223,65.141113 130.071991,68.490868 127.260376,71.859322   C124.613472,75.030449 128.325790,86.274353 132.597900,89.760208   C135.803009,92.375427 139.087112,94.896683 142.236649,97.576584   C149.427597,103.695267 148.581451,106.039536 144.755234,112.350838   C138.108780,123.314041 128.815033,129.647644 117.114983,133.571854   C113.529907,134.774292 110.403694,137.314957 106.837921,138.612122   C101.946121,140.391693 96.917999,142.102524 91.801537,142.893692   C81.726234,144.451630 73.032448,148.392944 66.536179,156.278671   C63.285271,160.224899 60.536346,164.656723 58.042816,169.137390   C54.591679,175.338821 49.421242,178.017532 42.553642,178.008820   C35.730309,178.000168 28.906944,178.014832 21.627445,178.031250   C18.908125,144.707748 28.793966,115.329132 49.903500,89.573189   C69.579376,65.566467 95.351990,52.118404 125.354950,46.151134   C130.170456,45.193378 135.213654,45.380424 140.610550,45.097717  z"
-      />
-      <g className="fill-black dark:fill-white">
-        <path d=" M901.044739,425.999878   C919.485718,427.998474 935.354919,434.289032 944.699707,450.720520   C949.495483,459.153320 950.662109,468.703339 949.259644,478.856171   C946.974243,495.400848 937.523254,505.586945 922.936646,512.671204   C933.363220,515.815063 941.442261,521.896851 947.145386,530.738647   C958.871887,548.918640 956.647583,580.028015 935.469482,594.330261   C924.158813,601.968750 911.978638,605.778381 898.664001,605.919250   C872.675049,606.194397 846.679749,605.852295 820.690125,606.094116   C813.326599,606.162537 805.698303,600.812866 805.005310,592.203247   C804.637512,587.634705 804.051392,583.061584 804.043762,578.489685   C803.970642,534.670166 803.963257,490.850403 804.066345,447.031036   C804.075012,443.341248 804.512024,439.488770 805.641663,436.000671   C807.752136,429.483948 814.375916,425.041412 821.362610,425.023773   C845.688110,424.962250 870.014038,424.975677 894.339539,425.043396   C896.422852,425.049164 898.504456,425.666534 901.044739,425.999878  M893.110474,574.998413   C894.183716,574.669678 895.238159,574.148621 896.333069,574.041931   C907.086304,572.994568 913.627563,567.557495 915.827576,556.660278   C917.533997,548.208374 913.136169,537.167114 905.144531,533.816895   C900.345520,531.805054 894.982544,530.389954 889.810181,530.167236   C877.668640,529.644653 865.489197,530.000000 853.325134,530.000000   C850.241516,530.000000 847.157959,530.000000 844.351074,530.000000   C844.351074,545.440369 844.351074,560.179382 844.351074,574.999939   C860.527405,574.999939 876.352783,574.999939 893.110474,574.998413  M911.000000,478.382660   C911.267151,467.259857 905.121826,459.677551 894.545227,458.034088   C890.853943,457.460541 887.097595,457.063904 883.366882,457.029449   C870.361694,456.909241 857.354797,456.982941 844.206177,456.982941   C844.206177,470.175201 844.206177,483.897400 844.206177,498.000000   C859.174377,498.000000 873.951965,498.374084 888.700256,497.877563   C899.816650,497.503326 909.390747,493.319000 911.000000,478.382660  z" />
-        <path d=" M1969.000000,560.000000   C1969.000000,569.997009 1968.894409,579.495483 1969.024536,588.990723   C1969.223145,603.481689 1954.466309,609.093201 1944.660645,607.600159   C1933.593384,605.914917 1927.825195,595.875366 1927.890259,585.545593   C1928.154053,543.575012 1928.000000,501.601807 1928.000000,459.000000   C1915.593018,459.000000 1903.335205,459.174652 1891.086914,458.908417   C1886.598145,458.810852 1882.015747,458.093475 1877.681396,456.908264   C1870.274414,454.882843 1867.520020,449.331146 1868.000122,438.861115   C1868.310913,432.083771 1872.268188,427.361603 1879.055664,426.041779   C1882.070435,425.455597 1885.174927,425.047852 1888.239502,425.043060   C1929.393066,424.978729 1970.547119,424.925964 2011.700317,425.074982   C2018.331177,425.098999 2025.355835,426.439880 2027.717529,433.681335   C2031.290405,444.636536 2030.049072,457.617889 2013.711670,457.939117   C1998.998291,458.228485 1984.274658,458.000183 1969.000000,458.000183   C1969.000000,492.656525 1969.000000,526.078247 1969.000000,560.000000  z" />
-        <path d=" M2477.765869,513.348389   C2469.297852,503.148468 2461.065674,493.223328 2452.760010,483.360352   C2446.750244,476.224091 2440.611328,469.196930 2434.592773,462.067932   C2431.868408,458.840637 2429.037598,455.656525 2426.709961,452.150879   C2421.802490,444.760345 2422.524170,433.859161 2431.491943,426.957153   C2437.152588,422.600311 2449.088867,422.186676 2454.650391,429.091400   C2463.649902,440.265015 2472.872803,451.259064 2482.032471,462.303162   C2489.438965,471.233429 2496.890869,480.125793 2504.470947,489.212860   C2513.100586,478.654144 2521.562012,468.322540 2529.999268,457.971527   C2537.924316,448.249268 2545.440918,438.158844 2553.856201,428.882782   C2560.616699,421.431091 2573.105225,422.318115 2578.959717,428.340942   C2587.581299,437.210571 2588.048340,446.590698 2580.730469,454.925446   C2570.722412,466.324310 2561.339844,478.270050 2551.577393,489.887482   C2543.681885,499.283508 2535.538086,508.472046 2527.708984,517.921997   C2526.297363,519.625854 2525.134277,522.145081 2525.115234,524.299194   C2524.925049,545.627930 2524.913574,566.959534 2525.028320,588.289429   C2525.114014,604.237610 2510.163574,610.042847 2498.480469,607.294800   C2488.446289,604.934631 2483.839111,596.359924 2483.941162,587.344482   C2484.141113,569.682434 2483.871826,552.015198 2484.059814,534.352844   C2484.141113,526.735229 2484.667969,519.185303 2477.765869,513.348389  z" />
-        <path d=" M1138.168213,354.444763   C1137.680176,304.485626 1137.186279,254.925644 1137.129028,205.365143   C1137.117554,195.392151 1140.182983,188.354126 1150.322754,184.939819   C1159.611816,181.812012 1168.279785,185.069031 1174.043457,192.746124   C1183.010864,204.690308 1192.361084,216.348190 1201.603027,228.084915   C1210.658813,239.585526 1219.809326,251.011780 1228.868164,262.510071   C1237.106812,272.967285 1245.281250,283.475067 1253.474731,293.967712   C1254.431030,295.192291 1255.328369,296.462982 1257.000000,298.721741   C1257.000000,296.055542 1257.000000,294.493561 1257.000000,292.931610   C1257.000000,262.605408 1257.267212,232.275406 1256.837646,201.955276   C1256.708252,192.815842 1265.968506,183.977829 1274.698853,183.750702   C1287.982910,183.405075 1295.034058,193.701462 1295.025269,202.051102   C1294.973999,250.872910 1295.008179,299.694794 1294.992798,348.516663   C1294.989868,357.771484 1290.157349,364.526978 1282.303955,367.019714   C1273.291260,369.880371 1265.162109,367.306580 1259.514648,360.045135   C1246.250244,342.989410 1232.828613,326.055908 1219.538208,309.020264   C1206.857544,292.766418 1194.276978,276.434753 1181.611572,260.169037   C1180.048218,258.161224 1178.234009,256.348755 1176.000000,253.844955   C1176.000000,255.948822 1176.000000,257.512146 1176.000000,259.075470   C1176.000000,287.402161 1176.057861,315.729095 1175.929321,344.055206   C1175.912598,347.743805 1175.084595,351.456787 1174.374146,355.108246   C1172.982544,362.262665 1164.277466,367.888489 1158.063721,368.185852   C1148.579590,368.639740 1141.159058,363.434937 1138.168213,354.444763  z" />
-        <path d=" M2337.772217,548.277222   C2332.006104,539.030151 2323.636230,535.803955 2313.511719,535.983398   C2308.201660,536.077515 2302.888428,536.000000 2297.240234,536.000000   C2297.153320,537.628174 2297.011475,539.063782 2297.010010,540.499451   C2296.993896,555.998657 2297.003174,571.497864 2296.998779,586.997070   C2296.996338,596.213989 2292.421875,604.722839 2285.666504,606.681824   C2276.361084,609.380188 2266.993408,609.307190 2260.771729,600.895630   C2257.922607,597.044006 2256.232910,591.395569 2256.198730,586.546631   C2255.867676,539.551392 2255.841064,492.552002 2256.104736,445.555878   C2256.176270,432.779816 2261.427246,424.943939 2276.846436,424.985077   C2298.845215,425.043762 2320.844727,424.910126 2342.842773,425.033386   C2356.275146,425.108643 2369.297607,427.393494 2381.180908,434.047119   C2393.207764,440.781403 2400.152100,451.280365 2402.999023,464.631042   C2405.765625,477.604309 2405.125977,490.209137 2400.132080,502.715271   C2395.069092,515.392944 2385.799805,523.483398 2373.805908,528.932068   C2370.481445,530.442383 2366.858398,531.295105 2363.058350,532.553528   C2375.896729,536.235474 2382.885254,546.252869 2389.742920,556.358337   C2394.989258,564.089111 2400.057129,571.940857 2405.223633,579.726013   C2411.030273,588.475647 2410.483887,598.510193 2403.456299,603.933838   C2394.431885,610.898987 2379.645752,610.464172 2372.218018,599.543884   C2360.707275,582.620972 2349.412842,565.551086 2337.772217,548.277222  M2357.891846,499.636017   C2365.239990,493.812683 2366.689453,485.657806 2365.874268,477.154602   C2365.036621,468.414398 2360.880127,461.391907 2351.702148,459.110413   C2346.715088,457.870789 2341.482666,457.181458 2336.343750,457.074921   C2323.006836,456.798462 2309.660156,456.982056 2296.227539,456.982056   C2296.227539,472.543884 2296.227539,488.603394 2296.227539,505.000000   C2309.234131,505.000000 2322.018799,504.832062 2334.796387,505.061554   C2342.757812,505.204529 2350.230713,503.816864 2357.891846,499.636017  z" />
-        <path d=" M1314.000000,455.013092   C1314.334229,448.941833 1314.486694,443.349487 1315.050659,437.798981   C1315.691895,431.486725 1323.501587,425.081726 1330.872803,425.046112   C1355.862427,424.925415 1380.874390,424.376556 1405.835693,425.226288   C1421.629761,425.763947 1437.007568,429.544250 1448.837891,441.403168   C1457.173462,449.759003 1460.759644,460.183594 1462.132080,471.815582   C1463.420654,482.735870 1462.048950,493.047791 1458.015625,503.139435   C1453.056152,515.547974 1443.731812,523.589172 1431.857178,528.930725   C1428.528198,530.428162 1424.906616,531.275391 1421.062256,532.538025   C1433.859131,536.269592 1440.870361,546.215881 1447.703613,556.302734   C1452.942139,564.035889 1458.093750,571.830688 1463.144897,579.687439   C1469.967041,590.298767 1468.851685,602.737122 1455.159424,607.211609   C1446.245605,610.124634 1436.588745,608.778015 1430.031860,599.202332   C1418.006348,581.640259 1405.881714,564.142029 1394.153198,546.382874   C1389.546997,539.408142 1382.957642,536.742188 1375.366821,536.113831   C1368.780762,535.568604 1362.113770,536.000366 1355.000000,536.000366   C1355.000000,553.237427 1355.095581,570.174133 1354.913940,587.107849   C1354.880371,590.243164 1354.206665,593.565247 1353.005493,596.459717   C1348.611816,607.047546 1337.035522,611.089050 1325.897339,606.454956   C1316.390015,602.499451 1314.109741,594.390686 1314.075317,585.459656   C1313.908447,542.143860 1314.000000,498.826996 1314.000000,455.013092  M1403.864990,457.339874   C1387.384399,456.703003 1370.903687,456.066162 1354.352051,455.426544   C1354.352051,472.573029 1354.352051,488.743958 1354.352051,504.999969   C1369.341553,504.999969 1383.980469,505.123047 1398.616455,504.963318   C1411.706055,504.820526 1420.846558,498.342926 1423.376099,485.946259   C1426.154785,472.328583 1420.867065,459.895233 1403.864990,457.339874  z" />
-        <path d=" M756.760254,221.264587   C763.567993,232.817993 768.015137,244.850510 768.927979,257.909637   C769.630798,267.965332 770.655762,278.153473 769.788879,288.130829   C768.115662,307.389923 762.266724,325.254700 749.266541,340.368195   C736.211853,355.545013 719.812073,363.823059 700.277649,367.103088   C677.788635,370.879211 656.231750,368.826569 636.022156,357.937195   C617.324219,347.862335 605.312622,332.302856 598.834900,311.989655   C594.712769,299.063324 592.722900,285.913300 593.154175,272.570892   C593.831665,251.612396 598.513794,231.532791 611.339478,214.342758   C622.293457,199.661346 637.344788,190.922318 654.840637,185.948730   C667.257629,182.418915 679.990051,182.087494 692.528870,183.195709   C719.625671,185.590561 741.898376,197.021561 756.760254,221.264587  M693.068604,335.976593   C703.593994,332.627441 712.626526,327.399139 719.099976,317.922089   C727.244263,305.998993 729.115295,292.455872 729.807800,278.743164   C730.200195,270.971649 729.173401,263.013885 727.853760,255.299408   C726.184265,245.539917 722.494995,236.501266 715.570435,229.006485   C705.673950,218.295013 693.426208,214.333786 679.087158,214.921783   C662.218018,215.613525 650.009644,223.330383 642.019531,237.978882   C637.202637,246.809891 635.988464,256.465820 634.760559,266.354095   C633.151672,279.311218 634.595581,291.830963 638.193542,303.988647   C642.972168,320.135620 653.266235,331.909607 670.340759,335.479401   C677.355652,336.946045 684.895752,335.901001 693.068604,335.976593  z" />
-        <path d=" M2087.343750,597.754333   C2065.382812,585.180481 2053.020020,566.093872 2048.172363,542.156555   C2046.140991,532.126160 2044.882324,521.674194 2045.235474,511.481995   C2045.942871,491.074463 2050.690918,471.512878 2063.096436,454.668701   C2073.995361,439.870453 2089.043457,431.099854 2106.446045,426.006531   C2119.299561,422.244507 2132.618164,421.874237 2145.613037,423.229156   C2177.228027,426.525543 2201.096191,441.871674 2214.150635,471.713074   C2219.174561,483.197113 2221.787598,495.346375 2222.035889,507.983276   C2222.329590,522.935242 2221.907227,537.761841 2217.028809,552.121094   C2212.922607,564.207153 2207.007568,575.346313 2197.975342,584.353882   C2185.537109,596.757996 2170.502686,604.549072 2152.929199,606.994141   C2147.772949,607.711609 2142.609131,608.847473 2137.438721,608.906250   C2120.043701,609.103943 2103.184570,606.463806 2087.343750,597.754333  M2086.032959,508.856750   C2086.688477,517.348694 2087.157715,525.860229 2088.045654,534.327820   C2089.204590,545.378418 2093.442627,555.401123 2101.072021,563.349976   C2112.016357,574.752441 2125.660645,578.559448 2141.420166,576.126038   C2154.372559,574.125977 2164.050293,567.918213 2171.448975,557.209900   C2179.480713,545.585876 2181.054199,532.281250 2181.799316,518.923279   C2182.232178,511.161133 2181.172852,503.200043 2179.861328,495.485413   C2178.203613,485.734314 2174.644531,476.632904 2167.695801,469.151550   C2158.096191,458.815826 2146.330078,454.713776 2132.287842,454.861420   C2112.051758,455.074127 2097.061768,466.909088 2091.143799,484.860535   C2088.694580,492.289185 2087.973145,500.287323 2086.032959,508.856750  z" />
-        <path d=" M710.562866,426.224243   C734.292236,433.450195 751.566833,447.831512 761.358337,470.152069   C765.113770,478.712891 767.320435,488.178558 768.771118,497.460358   C770.164429,506.374908 770.283508,515.612854 769.875305,524.659607   C768.942871,545.323975 762.967407,564.236450 749.243652,580.357056   C736.267517,595.599426 719.810791,603.824402 700.279480,607.102417   C677.794983,610.876038 656.246582,608.821228 636.032349,597.942017   C617.324097,587.873169 605.334045,572.303223 598.838989,552.010254   C594.703613,539.090088 592.725708,525.938782 593.153931,512.598389   C593.826599,491.643127 598.506897,471.570374 611.317505,454.372314   C622.262207,439.679199 637.319580,430.947632 654.797485,425.955841   C667.535767,422.317688 680.536499,422.759521 693.491028,423.142181   C699.080383,423.307251 704.624084,425.019958 710.562866,426.224243  M698.372803,574.002136   C711.948914,569.110962 720.473633,559.280334 724.960754,545.931702   C729.303040,533.014282 730.921082,519.825317 729.206421,506.060730   C728.117981,497.322632 727.088318,488.698364 723.234558,480.781189   C714.532959,462.904755 700.187500,454.594147 680.157959,454.869537   C660.229248,455.143616 645.081177,466.932037 639.031616,484.973541   C634.697449,497.899017 633.197083,511.078796 634.775574,524.848877   C635.932129,534.938538 637.797607,544.522949 642.216675,553.692688   C651.921265,573.830139 677.855286,581.839539 698.372803,574.002136  z" />
-        <path d=" M1099.646484,513.241577   C1111.545654,520.245422 1120.205933,529.086121 1123.519653,542.816956   C1127.478271,559.218811 1119.229614,583.155090 1105.526611,592.872559   C1092.572388,602.059082 1078.178223,606.546448 1062.406128,608.152344   C1042.225830,610.206909 1022.800964,607.644470 1003.857056,600.945129   C994.588928,597.667542 985.657166,593.574463 979.606873,585.158813   C973.788086,577.065063 978.536255,560.982849 988.833069,559.788025   C992.091125,559.409912 995.798767,560.233887 998.880493,561.516907   C1007.581177,565.139099 1015.804871,570.052856 1024.691040,573.051453   C1039.448730,578.031250 1054.679443,578.741028 1069.741455,573.941162   C1080.979370,570.360046 1086.666626,560.351135 1082.933228,550.740967   C1080.409058,544.243652 1074.291382,541.770935 1068.217896,540.069519   C1055.389282,536.476013 1042.260864,533.875610 1029.589600,529.823120   C1019.894958,526.722595 1010.358276,522.679138 1001.392029,517.862427   C989.873901,511.674835 983.924377,501.423431 981.816711,488.278748   C978.979431,470.584045 984.140869,455.662354 996.330200,443.436035   C1007.808899,431.922516 1022.358337,425.657562 1038.590454,424.089447   C1047.288086,423.249237 1056.135864,422.519165 1064.809204,423.166748   C1083.795166,424.584259 1101.731812,429.692749 1116.245972,442.859375   C1125.192627,450.975342 1122.850952,464.455444 1113.625610,470.594238   C1107.930542,474.383850 1101.479980,471.951935 1096.188721,468.537201   C1086.652466,462.383209 1076.847046,456.523041 1065.481201,455.415039   C1054.015747,454.297363 1042.304932,453.614014 1031.529053,459.850006   C1023.513611,464.488586 1018.134094,475.637207 1022.946838,483.790222   C1026.324707,489.512512 1031.919922,492.184052 1037.849487,493.947266   C1054.142334,498.792206 1070.557983,503.223846 1086.861206,508.034882   C1091.135986,509.296356 1095.178833,511.343414 1099.646484,513.241577  z" />
-        <path d=" M1207.028564,575.000000   C1228.687134,575.000000 1249.851196,575.065369 1271.014771,574.975098   C1283.322632,574.922546 1286.622314,582.621338 1286.805054,591.138123   C1286.998047,600.142883 1281.368652,606.059875 1270.853027,606.031494   C1238.690796,605.944824 1206.526733,605.824219 1174.367065,606.124634   C1166.408813,606.199036 1158.529297,600.774292 1157.983398,591.876709   C1157.733276,587.799377 1157.049805,583.731812 1157.043335,579.659058   C1156.973999,535.331665 1156.811157,491.002777 1157.124390,446.677490   C1157.209351,434.657806 1159.461914,429.048584 1170.911377,425.627136   C1172.440918,425.170044 1174.114990,425.026917 1175.722534,425.025146   C1206.885132,424.991211 1238.047852,425.062500 1269.210083,424.971100   C1282.814209,424.931213 1289.352539,434.614502 1285.921265,447.155212   C1284.507324,452.322693 1276.868042,456.986145 1269.938599,456.992401   C1247.441528,457.012726 1224.944580,457.000000 1202.447510,457.000000   C1200.808350,457.000000 1199.169189,457.000000 1197.266235,457.000000   C1197.266235,470.662018 1197.266235,484.066437 1197.266235,498.000000   C1201.854614,498.000000 1206.462158,498.000000 1211.069702,498.000000   C1228.067505,498.000000 1245.065918,497.915833 1262.062378,498.071838   C1265.410400,498.102570 1268.899536,498.616333 1272.061523,499.683624   C1278.148315,501.738068 1282.825684,509.536041 1281.962158,515.295532   C1280.564697,524.616272 1275.069092,529.950745 1266.474731,529.977356   C1244.477905,530.045471 1222.480591,530.000000 1200.483521,530.000061   C1199.504395,530.000061 1198.525146,530.000061 1197.278809,530.000061   C1197.278809,544.969543 1197.278809,559.707397 1197.278809,575.000000   C1200.319336,575.000000 1203.426636,575.000000 1207.028564,575.000000  z" />
-        <path d=" M980.000000,335.991394   C980.000000,292.665833 979.962097,249.838348 980.067078,207.011215   C980.076111,203.323029 980.521790,199.470917 981.657715,195.986099   C983.788696,189.448837 990.381653,185.034286 997.388062,185.022888   C1029.383545,184.970825 1061.379028,184.978668 1093.374634,185.016418   C1100.534058,185.024872 1107.084961,189.002304 1108.770630,194.462433   C1111.136475,202.126266 1110.090454,209.143982 1103.962158,213.282349   C1100.826050,215.400040 1096.554443,216.786255 1092.772095,216.840530   C1070.279907,217.163284 1047.780640,217.000000 1025.283813,217.000000   C1023.659424,217.000000 1022.035034,217.000000 1020.208374,217.000000   C1020.208374,230.780228 1020.208374,244.184021 1020.208374,258.000000   C1037.995117,258.000000 1055.611938,257.913055 1073.227173,258.045868   C1079.632446,258.094177 1086.059692,258.411560 1092.430176,259.063507   C1100.659180,259.905670 1105.432007,268.194153 1104.535645,276.283539   C1103.377075,286.739014 1095.406372,290.115631 1086.812622,290.050262   C1066.316650,289.894379 1045.818726,290.000000 1025.321533,290.000000   C1023.689148,290.000000 1022.056763,290.000000 1020.221680,290.000000   C1020.221680,305.086304 1020.221680,319.823334 1020.221680,335.000000   C1021.963135,335.000000 1023.579529,335.000000 1025.195923,335.000000   C1048.192749,335.000000 1071.189575,334.995300 1094.186401,335.002228   C1103.937866,335.005157 1110.086792,340.908081 1109.924072,350.803986   C1109.774658,359.882965 1104.599976,366.025970 1094.182129,366.013855   C1061.686646,365.976074 1029.191162,366.003845 996.695618,365.997498   C987.862122,365.995758 982.095886,361.049255 981.050598,352.199738   C980.435791,346.994263 980.333069,341.728302 980.000000,335.991394  z" />
-        <path d=" M1840.744141,543.334595   C1847.004395,556.933228 1852.989746,570.236084 1859.094116,583.484131   C1864.009155,594.151306 1856.761230,603.968811 1848.030640,606.352966   C1839.015991,608.814758 1830.241211,607.717285 1825.042725,596.890930   C1820.333740,587.083801 1816.052246,577.071472 1811.514893,567.000061   C1783.696655,567.000061 1755.882202,566.957153 1728.069092,567.129578   C1726.708374,567.138062 1724.765991,568.740295 1724.101318,570.094543   C1719.717163,579.027222 1715.380859,587.997986 1711.468506,597.144348   C1706.635254,608.444031 1695.131104,609.629089 1686.754883,605.595398   C1676.533081,600.672852 1672.601074,592.771423 1678.100342,580.816406   C1685.954712,563.741150 1694.415771,546.945984 1702.510010,529.979797   C1712.743652,508.529572 1722.871826,487.029114 1733.039551,465.547455   C1737.363403,456.412689 1741.773682,447.317017 1745.956177,438.117859   C1749.581787,430.143768 1755.225098,425.021332 1764.179443,423.990509   C1775.617676,422.673828 1784.810181,426.561005 1789.884277,436.960602   C1801.281860,460.320435 1812.043091,483.989929 1823.181396,507.477325   C1828.828857,519.386047 1834.732788,531.173157 1840.744141,543.334595  M1757.471924,495.020447   C1751.630371,508.162598 1745.788940,521.304810 1739.796875,534.785767   C1758.985718,534.785767 1777.741943,534.785767 1796.990479,534.785767   C1787.541992,513.510010 1778.132446,492.321747 1768.512573,470.660126   C1764.850952,478.902039 1761.406616,486.654968 1757.471924,495.020447  z" />
-        <path d=" M906.103638,186.975677   C928.509033,190.969360 944.776001,204.973846 949.036255,226.912811   C951.494141,239.570068 951.437622,252.081055 946.951355,264.227875   C941.054260,280.194763 929.578430,290.211395 913.257935,295.083649   C902.621277,298.259033 891.781555,298.908417 880.836914,298.982086   C869.046875,299.061432 857.255798,299.000000 845.000000,299.000000   C845.000000,315.922974 845.109009,332.526276 844.903015,349.125702   C844.864319,352.243683 844.057007,355.569885 842.756592,358.412109   C839.352234,365.852905 830.456238,369.826996 820.563110,367.877625   C812.318237,366.252991 806.049377,361.611603 805.028198,352.285950   C804.529663,347.733521 804.051331,343.148499 804.043762,338.577454   C803.970703,294.602570 803.807190,250.626114 804.126526,206.653442   C804.213623,194.656296 806.435608,189.067688 817.919556,185.619263   C819.447876,185.160324 821.123962,185.027557 822.732056,185.025055   C846.218628,184.988586 869.706055,184.905838 893.191284,185.087982   C897.353333,185.120255 901.506348,186.303207 906.103638,186.975677  M876.497559,267.000000   C880.826599,266.999969 885.164429,267.174011 889.482483,266.956635   C895.085754,266.674530 900.044067,264.649628 904.465332,261.068573   C912.789978,254.325974 913.111755,244.943375 911.684082,235.892990   C910.168457,226.284546 904.024536,219.687302 893.843872,217.994965   C890.487915,217.437134 887.072388,216.979782 883.679626,216.947754   C870.848450,216.826630 858.015442,216.899048 845.294434,216.899048   C845.294434,233.344574 845.294434,250.072403 845.294434,267.000000   C855.558105,267.000000 865.528259,267.000000 876.497559,267.000000  z" />
-        <path d=" M1603.005249,579.927856   C1600.586426,585.066711 1598.131470,589.781921 1595.835449,594.573242   C1591.112427,604.429321 1582.863403,608.415588 1572.480347,607.901672   C1562.888428,607.426880 1555.912842,602.792908 1551.875610,593.592773   C1546.910767,582.279297 1541.300049,571.250183 1536.012085,560.077576   C1523.304688,533.228088 1510.618774,506.368439 1497.932129,479.509216   C1493.269287,469.637390 1488.609009,459.764313 1483.983643,449.874878   C1479.739746,440.800903 1482.411499,431.929474 1490.934814,427.017975   C1501.862183,420.721191 1515.374268,422.910004 1521.125854,435.917755   C1533.707397,464.372162 1546.995728,492.513977 1559.976318,520.792236   C1564.847290,531.403564 1569.665527,542.039062 1574.766235,553.230591   C1583.648926,533.458008 1592.309570,513.988831 1601.136108,494.595184   C1607.717163,480.135529 1614.543457,465.787659 1621.196777,451.360535   C1623.611084,446.125183 1625.643188,440.706696 1628.203003,435.546967   C1631.510620,428.879395 1634.915527,422.765656 1649.557739,423.941345   C1655.568359,424.423981 1664.647095,430.276337 1665.561157,436.345215   C1666.253052,440.939331 1665.714355,446.351776 1663.860107,450.583801   C1656.901489,466.467010 1649.173828,482.013092 1641.743164,497.689667   C1629.432373,523.661987 1617.127563,549.637146 1604.824585,575.613098   C1604.209473,576.911865 1603.634766,578.229736 1603.005249,579.927856  z" />
-      </g>
-    </svg>
-  );
-}
-
-export default Logo;
+export const Logo = (props: DetailedHTMLProps<SVGAttributes<SVGSVGElement>, SVGSVGElement>) => (
+  <svg {...props} version="1.1" viewBox="0 0 2650 700" xmlns="http://www.w3.org/2000/svg">
+    <path
+      fill="#00BAFF"
+      d=" M28.167587,354.868042   C35.868050,322.277985 49.972065,292.650879 69.805077,265.758789   C82.927628,247.965607 98.131035,232.223145 115.652962,218.562561   C131.263702,206.392014 147.873367,196.072845 165.742294,187.717606   C184.432465,178.978363 203.937958,172.768509 224.240341,168.938583   C238.493698,166.249786 252.793274,164.587326 267.710114,165.536499   C268.075043,170.383698 267.933899,174.714981 268.019684,179.041748   C268.273804,191.860657 264.453552,203.602448 258.894226,214.946747   C258.135773,216.494446 257.880524,218.428650 257.876312,220.185059   C257.814423,246.004288 257.708099,271.824829 257.933441,297.642242   C257.983398,303.366547 258.859131,309.169708 260.074432,314.780670   C263.005707,328.314270 269.041656,341.032593 270.181396,355.093872   C270.986725,365.029114 271.886536,374.602692 268.133484,384.404755   C266.274048,389.261200 263.411102,393.162720 261.031525,397.524963   C255.011490,408.560883 248.986923,419.616943 243.628555,430.978302   C241.860840,434.726440 240.883194,439.684753 241.686218,443.656708   C244.532364,457.734375 253.752426,466.738342 266.264832,473.438049   C276.949280,479.158997 288.487488,482.043427 299.989655,483.639587   C315.513580,485.793793 331.379211,485.832886 347.107422,486.024170   C355.194855,486.122559 363.261627,483.611389 368.446320,477.006653   C374.744568,468.983368 380.284119,460.313202 385.602966,451.588226   C390.050781,444.292114 395.935120,438.959991 404.177917,437.266785   C418.220306,434.382324 432.219757,435.305145 445.727051,440.356445   C455.229767,443.910187 464.810333,446.572266 475.095184,446.076324   C476.955322,445.986633 478.855133,446.934631 480.751221,447.334534   C486.628845,448.574188 492.485016,450.578461 498.402924,450.789642   C505.184998,451.031708 512.024231,449.671265 518.838684,449.006989   C517.117126,458.574463 515.878845,468.260193 513.572021,477.684418   C508.978210,496.452148 502.054230,514.397644 493.125244,531.572571   C484.997528,547.206177 475.573669,561.965027 464.101807,575.349854   C458.206146,582.228638 452.374390,589.230713 445.861664,595.499573   C423.955292,616.585693 399.124878,633.301758 371.161560,645.537109   C352.944733,653.507935 334.078583,659.057739 314.699463,662.859497   C297.475037,666.238586 279.938934,666.220947 262.419495,666.052856   C241.028198,665.847534 220.311996,661.983154 199.817825,656.092712   C172.245255,648.167725 146.857834,635.667358 123.902580,618.753784   C110.840797,609.129700 99.328812,597.402283 87.155746,585.961548   C88.847534,579.962341 90.492218,574.620483 92.163025,569.286804   C92.740837,567.442200 93.498398,565.647400 93.977486,563.779480   C96.512589,553.895264 98.231499,543.721008 101.700089,534.180542   C104.903587,525.369263 101.033112,513.152283 92.990562,507.287781   C87.158600,503.035187 79.954857,497.365906 82.046188,487.770020   C82.629921,485.091614 83.486923,482.393433 84.713257,479.952881   C86.567871,476.262024 89.483871,473.011475 90.840782,469.181274   C94.678528,458.348389 97.319405,447.062317 101.582314,436.415405   C110.231949,414.812439 96.112907,391.530853 81.856262,383.153473   C64.435081,372.916565 46.104771,364.226807 28.167587,354.868042  M231.298508,607.789551   C227.994736,610.190979 224.690979,612.592407 220.566452,615.590393   C221.354904,617.586487 222.064163,620.289795 223.432648,622.603088   C224.940567,625.152100 227.074356,627.327148 228.897736,629.694214   C230.708786,632.045227 232.617111,633.427368 235.891266,634.295715   C242.986343,636.177612 250.636139,637.540588 255.544830,644.366455   C257.080688,646.502197 260.127258,648.607910 262.636200,648.809265   C275.064758,649.806519 287.159760,647.564087 298.989716,643.728210   C305.123199,641.739380 311.248779,638.660461 317.493896,638.341003   C326.039612,637.903870 334.111908,636.031921 341.613861,632.671326   C346.017456,630.698730 351.531525,626.555725 352.475586,622.452515   C354.609894,613.176086 353.081573,603.602905 345.831482,596.129883   C341.771301,591.944702 337.747742,587.698669 333.385284,583.844910   C331.718292,582.372314 328.811249,580.707092 327.052490,581.202087   C317.881042,583.783142 308.378662,586.011597 299.970886,590.288330   C287.605438,596.578064 276.073639,600.619629 262.413971,593.629211   C254.923157,589.795898 247.061752,590.930542 241.235306,598.688599   C238.724380,602.032043 235.023148,604.481445 231.298508,607.789551 z"
+    />
+    <path
+      fill="#00B501"
+      d=" M231.580597,607.563232   C235.023148,604.481445 238.724380,602.032043 241.235306,598.688599   C247.061752,590.930542 254.923157,589.795898 262.413971,593.629211   C276.073639,600.619629 287.605438,596.578064 299.970886,590.288330   C308.378662,586.011597 317.881042,583.783142 327.052490,581.202087   C328.811249,580.707092 331.718292,582.372314 333.385284,583.844910   C337.747742,587.698669 341.771301,591.944702 345.831482,596.129883   C353.081573,603.602905 354.609894,613.176086 352.475586,622.452515   C351.531525,626.555725 346.017456,630.698730 341.613861,632.671326   C334.111908,636.031921 326.039612,637.903870 317.493896,638.341003   C311.248779,638.660461 305.123199,641.739380 298.989716,643.728210   C287.159760,647.564087 275.064758,649.806519 262.636200,648.809265   C260.127258,648.607910 257.080688,646.502197 255.544830,644.366455   C250.636139,637.540588 242.986343,636.177612 235.891266,634.295715   C232.617111,633.427368 230.708786,632.045227 228.897736,629.694214   C227.074356,627.327148 224.940567,625.152100 223.432648,622.603088   C222.064163,620.289795 221.354904,617.586487 220.566452,615.590393   C224.690979,612.592407 227.994736,610.190979 231.580597,607.563232  z"
+    />
+    <path
+      fill="#00BAFB"
+      d=" M519.370117,446.942352   C518.667053,446.479004 518.004944,445.764435 518.050110,445.097687   C518.615417,436.739807 519.285889,428.389099 519.921814,420.035980   C520.035339,418.545013 520.114868,417.051453 520.525513,415.319275   C520.473877,425.624054 520.106873,436.168701 519.370117,446.942352  z"
+    />
+    <path
+      fill="#666666"
+      d=" M265.009949,128.533661   C263.641663,133.968658 259.973389,131.882767 256.944641,131.403793   C252.602493,130.717087 248.323746,129.553268 243.965057,129.063889   C236.171356,128.188858 228.920319,126.208847 224.187042,119.453926   C222.951370,117.690460 222.927948,114.971542 222.672226,112.655617   C222.385757,110.061180 222.758636,107.376442 222.322601,104.820038   C221.630081,100.759674 220.155182,98.134880 214.961990,100.335991   C210.931198,102.044434 205.621414,98.917557 203.173035,94.262817   C200.985229,90.103500 201.370758,86.327354 205.277313,83.197380   C210.683334,78.866005 216.012924,74.439224 221.725922,70.021103   C232.301041,76.411461 240.159500,85.230125 247.311707,94.789009   C254.948608,104.995659 260.890991,116.084435 265.009949,128.533661  z"
+    />
+    <path
+      fill="#00B500"
+      d=" M518.900635,448.705261   C512.024231,449.671265 505.184998,451.031708 498.402924,450.789642   C492.485016,450.578461 486.628845,448.574188 480.751221,447.334534   C478.855133,446.934631 476.955322,445.986633 475.095184,446.076324   C464.810333,446.572266 455.229767,443.910187 445.727051,440.356445   C432.219757,435.305145 418.220306,434.382324 404.177917,437.266785   C395.935120,438.959991 390.050781,444.292114 385.602966,451.588226   C380.284119,460.313202 374.744568,468.983368 368.446320,477.006653   C363.261627,483.611389 355.194855,486.122559 347.107422,486.024170   C331.379211,485.832886 315.513580,485.793793 299.989655,483.639587   C288.487488,482.043427 276.949280,479.158997 266.264832,473.438049   C253.752426,466.738342 244.532364,457.734375 241.686218,443.656708   C240.883194,439.684753 241.860840,434.726440 243.628555,430.978302   C248.986923,419.616943 255.011490,408.560883 261.031525,397.524963   C263.411102,393.162720 266.274048,389.261200 268.133484,384.404755   C271.886536,374.602692 270.986725,365.029114 270.181396,355.093872   C269.041656,341.032593 263.005707,328.314270 260.074432,314.780670   C258.859131,309.169708 257.983398,303.366547 257.933441,297.642242   C257.708099,271.824829 257.814423,246.004288 257.876312,220.185059   C257.880524,218.428650 258.135773,216.494446 258.894226,214.946747   C264.453552,203.602448 268.273804,191.860657 268.019684,179.041748   C267.933899,174.714981 268.075043,170.383698 268.191406,165.603180   C293.549225,163.907318 318.108063,167.841995 342.325745,174.915939   C371.543732,183.450500 398.177368,197.115952 422.409607,215.296295   C439.351318,228.006866 454.432098,242.741333 467.451202,259.639740   C479.057983,274.704895 489.272064,290.538666 497.279541,307.743835   C507.291199,329.255249 514.930969,351.598694 517.908325,375.224213   C519.533386,388.119415 519.982788,401.162750 520.586182,414.849182   C520.114868,417.051453 520.035339,418.545013 519.921814,420.035980   C519.285889,428.389099 518.615417,436.739807 518.050110,445.097687   C518.004944,445.764435 518.667053,446.479004 519.000427,447.171356   C518.987793,447.582062 518.975159,447.992798 518.900635,448.705261  z"
+    />
+    <path
+      fill="#00B500"
+      d=" M28.086550,355.266937   C46.104771,364.226807 64.435081,372.916565 81.856262,383.153473   C96.112907,391.530853 110.231949,414.812439 101.582314,436.415405   C97.319405,447.062317 94.678528,458.348389 90.840782,469.181274   C89.483871,473.011475 86.567871,476.262024 84.713257,479.952881   C83.486923,482.393433 82.629921,485.091614 82.046188,487.770020   C79.954857,497.365906 87.158600,503.035187 92.990562,507.287781   C101.033112,513.152283 104.903587,525.369263 101.700089,534.180542   C98.231499,543.721008 96.512589,553.895264 93.977486,563.779480   C93.498398,565.647400 92.740837,567.442200 92.163025,569.286804   C90.492218,574.620483 88.847534,579.962341 87.013214,585.651123   C77.587128,573.867859 67.571922,562.236206 59.256271,549.495239   C42.765011,524.227905 32.403427,496.351013 25.836536,466.868835   C20.668997,443.669098 20.105732,420.240967 21.298513,396.829681   C22.001575,383.030304 25.670464,369.382080 28.086550,355.266937  z"
+    />
+    <path
+      fill="#7E7E7E"
+      d=" M141.069504,45.157181   C170.998398,43.143059 197.655746,52.043030 221.583542,69.703674   C216.012924,74.439224 210.683334,78.866005 205.277313,83.197380   C201.370758,86.327354 200.985229,90.103500 203.173035,94.262817   C205.621414,98.917557 210.931198,102.044434 214.961990,100.335991   C220.155182,98.134880 221.630081,100.759674 222.322601,104.820038   C222.758636,107.376442 222.385757,110.061180 222.672226,112.655617   C222.927948,114.971542 222.951370,117.690460 224.187042,119.453926   C228.920319,126.208847 236.171356,128.188858 243.965057,129.063889   C248.323746,129.553268 252.602493,130.717087 256.944641,131.403793   C259.973389,131.882767 263.641663,133.968658 265.020172,128.913849   C269.571960,138.627701 270.715302,149.070221 272.108276,159.857117   C230.363998,158.656036 191.489380,168.150406 154.878464,186.681641   C117.979553,205.358643 87.864510,232.118225 63.655178,265.882141   C48.646038,252.609314 37.576401,236.751312 29.994730,218.412704   C24.765434,205.764023 21.450434,192.673157 21.561848,178.473450   C28.906944,178.014832 35.730309,178.000168 42.553642,178.008820   C49.421242,178.017532 54.591679,175.338821 58.042816,169.137390   C60.536346,164.656723 63.285271,160.224899 66.536179,156.278671   C73.032448,148.392944 81.726234,144.451630 91.801537,142.893692   C96.917999,142.102524 101.946121,140.391693 106.837921,138.612122   C110.403694,137.314957 113.529907,134.774292 117.114983,133.571854   C128.815033,129.647644 138.108780,123.314041 144.755234,112.350838   C148.581451,106.039536 149.427597,103.695267 142.236649,97.576584   C139.087112,94.896683 135.803009,92.375427 132.597900,89.760208   C128.325790,86.274353 124.613472,75.030449 127.260376,71.859322   C130.071991,68.490868 133.400223,65.141113 137.208694,63.100632   C145.905457,58.441132 147.396515,52.827572 141.069504,45.157181  z"
+    />
+    <path
+      fill="#666666"
+      d=" M140.610550,45.097717   C147.396515,52.827572 145.905457,58.441132 137.208694,63.100632   C133.400223,65.141113 130.071991,68.490868 127.260376,71.859322   C124.613472,75.030449 128.325790,86.274353 132.597900,89.760208   C135.803009,92.375427 139.087112,94.896683 142.236649,97.576584   C149.427597,103.695267 148.581451,106.039536 144.755234,112.350838   C138.108780,123.314041 128.815033,129.647644 117.114983,133.571854   C113.529907,134.774292 110.403694,137.314957 106.837921,138.612122   C101.946121,140.391693 96.917999,142.102524 91.801537,142.893692   C81.726234,144.451630 73.032448,148.392944 66.536179,156.278671   C63.285271,160.224899 60.536346,164.656723 58.042816,169.137390   C54.591679,175.338821 49.421242,178.017532 42.553642,178.008820   C35.730309,178.000168 28.906944,178.014832 21.627445,178.031250   C18.908125,144.707748 28.793966,115.329132 49.903500,89.573189   C69.579376,65.566467 95.351990,52.118404 125.354950,46.151134   C130.170456,45.193378 135.213654,45.380424 140.610550,45.097717  z"
+    />
+    <g className="fill-black dark:fill-white">
+      <path d=" M901.044739,425.999878   C919.485718,427.998474 935.354919,434.289032 944.699707,450.720520   C949.495483,459.153320 950.662109,468.703339 949.259644,478.856171   C946.974243,495.400848 937.523254,505.586945 922.936646,512.671204   C933.363220,515.815063 941.442261,521.896851 947.145386,530.738647   C958.871887,548.918640 956.647583,580.028015 935.469482,594.330261   C924.158813,601.968750 911.978638,605.778381 898.664001,605.919250   C872.675049,606.194397 846.679749,605.852295 820.690125,606.094116   C813.326599,606.162537 805.698303,600.812866 805.005310,592.203247   C804.637512,587.634705 804.051392,583.061584 804.043762,578.489685   C803.970642,534.670166 803.963257,490.850403 804.066345,447.031036   C804.075012,443.341248 804.512024,439.488770 805.641663,436.000671   C807.752136,429.483948 814.375916,425.041412 821.362610,425.023773   C845.688110,424.962250 870.014038,424.975677 894.339539,425.043396   C896.422852,425.049164 898.504456,425.666534 901.044739,425.999878  M893.110474,574.998413   C894.183716,574.669678 895.238159,574.148621 896.333069,574.041931   C907.086304,572.994568 913.627563,567.557495 915.827576,556.660278   C917.533997,548.208374 913.136169,537.167114 905.144531,533.816895   C900.345520,531.805054 894.982544,530.389954 889.810181,530.167236   C877.668640,529.644653 865.489197,530.000000 853.325134,530.000000   C850.241516,530.000000 847.157959,530.000000 844.351074,530.000000   C844.351074,545.440369 844.351074,560.179382 844.351074,574.999939   C860.527405,574.999939 876.352783,574.999939 893.110474,574.998413  M911.000000,478.382660   C911.267151,467.259857 905.121826,459.677551 894.545227,458.034088   C890.853943,457.460541 887.097595,457.063904 883.366882,457.029449   C870.361694,456.909241 857.354797,456.982941 844.206177,456.982941   C844.206177,470.175201 844.206177,483.897400 844.206177,498.000000   C859.174377,498.000000 873.951965,498.374084 888.700256,497.877563   C899.816650,497.503326 909.390747,493.319000 911.000000,478.382660  z" />
+      <path d=" M1969.000000,560.000000   C1969.000000,569.997009 1968.894409,579.495483 1969.024536,588.990723   C1969.223145,603.481689 1954.466309,609.093201 1944.660645,607.600159   C1933.593384,605.914917 1927.825195,595.875366 1927.890259,585.545593   C1928.154053,543.575012 1928.000000,501.601807 1928.000000,459.000000   C1915.593018,459.000000 1903.335205,459.174652 1891.086914,458.908417   C1886.598145,458.810852 1882.015747,458.093475 1877.681396,456.908264   C1870.274414,454.882843 1867.520020,449.331146 1868.000122,438.861115   C1868.310913,432.083771 1872.268188,427.361603 1879.055664,426.041779   C1882.070435,425.455597 1885.174927,425.047852 1888.239502,425.043060   C1929.393066,424.978729 1970.547119,424.925964 2011.700317,425.074982   C2018.331177,425.098999 2025.355835,426.439880 2027.717529,433.681335   C2031.290405,444.636536 2030.049072,457.617889 2013.711670,457.939117   C1998.998291,458.228485 1984.274658,458.000183 1969.000000,458.000183   C1969.000000,492.656525 1969.000000,526.078247 1969.000000,560.000000  z" />
+      <path d=" M2477.765869,513.348389   C2469.297852,503.148468 2461.065674,493.223328 2452.760010,483.360352   C2446.750244,476.224091 2440.611328,469.196930 2434.592773,462.067932   C2431.868408,458.840637 2429.037598,455.656525 2426.709961,452.150879   C2421.802490,444.760345 2422.524170,433.859161 2431.491943,426.957153   C2437.152588,422.600311 2449.088867,422.186676 2454.650391,429.091400   C2463.649902,440.265015 2472.872803,451.259064 2482.032471,462.303162   C2489.438965,471.233429 2496.890869,480.125793 2504.470947,489.212860   C2513.100586,478.654144 2521.562012,468.322540 2529.999268,457.971527   C2537.924316,448.249268 2545.440918,438.158844 2553.856201,428.882782   C2560.616699,421.431091 2573.105225,422.318115 2578.959717,428.340942   C2587.581299,437.210571 2588.048340,446.590698 2580.730469,454.925446   C2570.722412,466.324310 2561.339844,478.270050 2551.577393,489.887482   C2543.681885,499.283508 2535.538086,508.472046 2527.708984,517.921997   C2526.297363,519.625854 2525.134277,522.145081 2525.115234,524.299194   C2524.925049,545.627930 2524.913574,566.959534 2525.028320,588.289429   C2525.114014,604.237610 2510.163574,610.042847 2498.480469,607.294800   C2488.446289,604.934631 2483.839111,596.359924 2483.941162,587.344482   C2484.141113,569.682434 2483.871826,552.015198 2484.059814,534.352844   C2484.141113,526.735229 2484.667969,519.185303 2477.765869,513.348389  z" />
+      <path d=" M1138.168213,354.444763   C1137.680176,304.485626 1137.186279,254.925644 1137.129028,205.365143   C1137.117554,195.392151 1140.182983,188.354126 1150.322754,184.939819   C1159.611816,181.812012 1168.279785,185.069031 1174.043457,192.746124   C1183.010864,204.690308 1192.361084,216.348190 1201.603027,228.084915   C1210.658813,239.585526 1219.809326,251.011780 1228.868164,262.510071   C1237.106812,272.967285 1245.281250,283.475067 1253.474731,293.967712   C1254.431030,295.192291 1255.328369,296.462982 1257.000000,298.721741   C1257.000000,296.055542 1257.000000,294.493561 1257.000000,292.931610   C1257.000000,262.605408 1257.267212,232.275406 1256.837646,201.955276   C1256.708252,192.815842 1265.968506,183.977829 1274.698853,183.750702   C1287.982910,183.405075 1295.034058,193.701462 1295.025269,202.051102   C1294.973999,250.872910 1295.008179,299.694794 1294.992798,348.516663   C1294.989868,357.771484 1290.157349,364.526978 1282.303955,367.019714   C1273.291260,369.880371 1265.162109,367.306580 1259.514648,360.045135   C1246.250244,342.989410 1232.828613,326.055908 1219.538208,309.020264   C1206.857544,292.766418 1194.276978,276.434753 1181.611572,260.169037   C1180.048218,258.161224 1178.234009,256.348755 1176.000000,253.844955   C1176.000000,255.948822 1176.000000,257.512146 1176.000000,259.075470   C1176.000000,287.402161 1176.057861,315.729095 1175.929321,344.055206   C1175.912598,347.743805 1175.084595,351.456787 1174.374146,355.108246   C1172.982544,362.262665 1164.277466,367.888489 1158.063721,368.185852   C1148.579590,368.639740 1141.159058,363.434937 1138.168213,354.444763  z" />
+      <path d=" M2337.772217,548.277222   C2332.006104,539.030151 2323.636230,535.803955 2313.511719,535.983398   C2308.201660,536.077515 2302.888428,536.000000 2297.240234,536.000000   C2297.153320,537.628174 2297.011475,539.063782 2297.010010,540.499451   C2296.993896,555.998657 2297.003174,571.497864 2296.998779,586.997070   C2296.996338,596.213989 2292.421875,604.722839 2285.666504,606.681824   C2276.361084,609.380188 2266.993408,609.307190 2260.771729,600.895630   C2257.922607,597.044006 2256.232910,591.395569 2256.198730,586.546631   C2255.867676,539.551392 2255.841064,492.552002 2256.104736,445.555878   C2256.176270,432.779816 2261.427246,424.943939 2276.846436,424.985077   C2298.845215,425.043762 2320.844727,424.910126 2342.842773,425.033386   C2356.275146,425.108643 2369.297607,427.393494 2381.180908,434.047119   C2393.207764,440.781403 2400.152100,451.280365 2402.999023,464.631042   C2405.765625,477.604309 2405.125977,490.209137 2400.132080,502.715271   C2395.069092,515.392944 2385.799805,523.483398 2373.805908,528.932068   C2370.481445,530.442383 2366.858398,531.295105 2363.058350,532.553528   C2375.896729,536.235474 2382.885254,546.252869 2389.742920,556.358337   C2394.989258,564.089111 2400.057129,571.940857 2405.223633,579.726013   C2411.030273,588.475647 2410.483887,598.510193 2403.456299,603.933838   C2394.431885,610.898987 2379.645752,610.464172 2372.218018,599.543884   C2360.707275,582.620972 2349.412842,565.551086 2337.772217,548.277222  M2357.891846,499.636017   C2365.239990,493.812683 2366.689453,485.657806 2365.874268,477.154602   C2365.036621,468.414398 2360.880127,461.391907 2351.702148,459.110413   C2346.715088,457.870789 2341.482666,457.181458 2336.343750,457.074921   C2323.006836,456.798462 2309.660156,456.982056 2296.227539,456.982056   C2296.227539,472.543884 2296.227539,488.603394 2296.227539,505.000000   C2309.234131,505.000000 2322.018799,504.832062 2334.796387,505.061554   C2342.757812,505.204529 2350.230713,503.816864 2357.891846,499.636017  z" />
+      <path d=" M1314.000000,455.013092   C1314.334229,448.941833 1314.486694,443.349487 1315.050659,437.798981   C1315.691895,431.486725 1323.501587,425.081726 1330.872803,425.046112   C1355.862427,424.925415 1380.874390,424.376556 1405.835693,425.226288   C1421.629761,425.763947 1437.007568,429.544250 1448.837891,441.403168   C1457.173462,449.759003 1460.759644,460.183594 1462.132080,471.815582   C1463.420654,482.735870 1462.048950,493.047791 1458.015625,503.139435   C1453.056152,515.547974 1443.731812,523.589172 1431.857178,528.930725   C1428.528198,530.428162 1424.906616,531.275391 1421.062256,532.538025   C1433.859131,536.269592 1440.870361,546.215881 1447.703613,556.302734   C1452.942139,564.035889 1458.093750,571.830688 1463.144897,579.687439   C1469.967041,590.298767 1468.851685,602.737122 1455.159424,607.211609   C1446.245605,610.124634 1436.588745,608.778015 1430.031860,599.202332   C1418.006348,581.640259 1405.881714,564.142029 1394.153198,546.382874   C1389.546997,539.408142 1382.957642,536.742188 1375.366821,536.113831   C1368.780762,535.568604 1362.113770,536.000366 1355.000000,536.000366   C1355.000000,553.237427 1355.095581,570.174133 1354.913940,587.107849   C1354.880371,590.243164 1354.206665,593.565247 1353.005493,596.459717   C1348.611816,607.047546 1337.035522,611.089050 1325.897339,606.454956   C1316.390015,602.499451 1314.109741,594.390686 1314.075317,585.459656   C1313.908447,542.143860 1314.000000,498.826996 1314.000000,455.013092  M1403.864990,457.339874   C1387.384399,456.703003 1370.903687,456.066162 1354.352051,455.426544   C1354.352051,472.573029 1354.352051,488.743958 1354.352051,504.999969   C1369.341553,504.999969 1383.980469,505.123047 1398.616455,504.963318   C1411.706055,504.820526 1420.846558,498.342926 1423.376099,485.946259   C1426.154785,472.328583 1420.867065,459.895233 1403.864990,457.339874  z" />
+      <path d=" M756.760254,221.264587   C763.567993,232.817993 768.015137,244.850510 768.927979,257.909637   C769.630798,267.965332 770.655762,278.153473 769.788879,288.130829   C768.115662,307.389923 762.266724,325.254700 749.266541,340.368195   C736.211853,355.545013 719.812073,363.823059 700.277649,367.103088   C677.788635,370.879211 656.231750,368.826569 636.022156,357.937195   C617.324219,347.862335 605.312622,332.302856 598.834900,311.989655   C594.712769,299.063324 592.722900,285.913300 593.154175,272.570892   C593.831665,251.612396 598.513794,231.532791 611.339478,214.342758   C622.293457,199.661346 637.344788,190.922318 654.840637,185.948730   C667.257629,182.418915 679.990051,182.087494 692.528870,183.195709   C719.625671,185.590561 741.898376,197.021561 756.760254,221.264587  M693.068604,335.976593   C703.593994,332.627441 712.626526,327.399139 719.099976,317.922089   C727.244263,305.998993 729.115295,292.455872 729.807800,278.743164   C730.200195,270.971649 729.173401,263.013885 727.853760,255.299408   C726.184265,245.539917 722.494995,236.501266 715.570435,229.006485   C705.673950,218.295013 693.426208,214.333786 679.087158,214.921783   C662.218018,215.613525 650.009644,223.330383 642.019531,237.978882   C637.202637,246.809891 635.988464,256.465820 634.760559,266.354095   C633.151672,279.311218 634.595581,291.830963 638.193542,303.988647   C642.972168,320.135620 653.266235,331.909607 670.340759,335.479401   C677.355652,336.946045 684.895752,335.901001 693.068604,335.976593  z" />
+      <path d=" M2087.343750,597.754333   C2065.382812,585.180481 2053.020020,566.093872 2048.172363,542.156555   C2046.140991,532.126160 2044.882324,521.674194 2045.235474,511.481995   C2045.942871,491.074463 2050.690918,471.512878 2063.096436,454.668701   C2073.995361,439.870453 2089.043457,431.099854 2106.446045,426.006531   C2119.299561,422.244507 2132.618164,421.874237 2145.613037,423.229156   C2177.228027,426.525543 2201.096191,441.871674 2214.150635,471.713074   C2219.174561,483.197113 2221.787598,495.346375 2222.035889,507.983276   C2222.329590,522.935242 2221.907227,537.761841 2217.028809,552.121094   C2212.922607,564.207153 2207.007568,575.346313 2197.975342,584.353882   C2185.537109,596.757996 2170.502686,604.549072 2152.929199,606.994141   C2147.772949,607.711609 2142.609131,608.847473 2137.438721,608.906250   C2120.043701,609.103943 2103.184570,606.463806 2087.343750,597.754333  M2086.032959,508.856750   C2086.688477,517.348694 2087.157715,525.860229 2088.045654,534.327820   C2089.204590,545.378418 2093.442627,555.401123 2101.072021,563.349976   C2112.016357,574.752441 2125.660645,578.559448 2141.420166,576.126038   C2154.372559,574.125977 2164.050293,567.918213 2171.448975,557.209900   C2179.480713,545.585876 2181.054199,532.281250 2181.799316,518.923279   C2182.232178,511.161133 2181.172852,503.200043 2179.861328,495.485413   C2178.203613,485.734314 2174.644531,476.632904 2167.695801,469.151550   C2158.096191,458.815826 2146.330078,454.713776 2132.287842,454.861420   C2112.051758,455.074127 2097.061768,466.909088 2091.143799,484.860535   C2088.694580,492.289185 2087.973145,500.287323 2086.032959,508.856750  z" />
+      <path d=" M710.562866,426.224243   C734.292236,433.450195 751.566833,447.831512 761.358337,470.152069   C765.113770,478.712891 767.320435,488.178558 768.771118,497.460358   C770.164429,506.374908 770.283508,515.612854 769.875305,524.659607   C768.942871,545.323975 762.967407,564.236450 749.243652,580.357056   C736.267517,595.599426 719.810791,603.824402 700.279480,607.102417   C677.794983,610.876038 656.246582,608.821228 636.032349,597.942017   C617.324097,587.873169 605.334045,572.303223 598.838989,552.010254   C594.703613,539.090088 592.725708,525.938782 593.153931,512.598389   C593.826599,491.643127 598.506897,471.570374 611.317505,454.372314   C622.262207,439.679199 637.319580,430.947632 654.797485,425.955841   C667.535767,422.317688 680.536499,422.759521 693.491028,423.142181   C699.080383,423.307251 704.624084,425.019958 710.562866,426.224243  M698.372803,574.002136   C711.948914,569.110962 720.473633,559.280334 724.960754,545.931702   C729.303040,533.014282 730.921082,519.825317 729.206421,506.060730   C728.117981,497.322632 727.088318,488.698364 723.234558,480.781189   C714.532959,462.904755 700.187500,454.594147 680.157959,454.869537   C660.229248,455.143616 645.081177,466.932037 639.031616,484.973541   C634.697449,497.899017 633.197083,511.078796 634.775574,524.848877   C635.932129,534.938538 637.797607,544.522949 642.216675,553.692688   C651.921265,573.830139 677.855286,581.839539 698.372803,574.002136  z" />
+      <path d=" M1099.646484,513.241577   C1111.545654,520.245422 1120.205933,529.086121 1123.519653,542.816956   C1127.478271,559.218811 1119.229614,583.155090 1105.526611,592.872559   C1092.572388,602.059082 1078.178223,606.546448 1062.406128,608.152344   C1042.225830,610.206909 1022.800964,607.644470 1003.857056,600.945129   C994.588928,597.667542 985.657166,593.574463 979.606873,585.158813   C973.788086,577.065063 978.536255,560.982849 988.833069,559.788025   C992.091125,559.409912 995.798767,560.233887 998.880493,561.516907   C1007.581177,565.139099 1015.804871,570.052856 1024.691040,573.051453   C1039.448730,578.031250 1054.679443,578.741028 1069.741455,573.941162   C1080.979370,570.360046 1086.666626,560.351135 1082.933228,550.740967   C1080.409058,544.243652 1074.291382,541.770935 1068.217896,540.069519   C1055.389282,536.476013 1042.260864,533.875610 1029.589600,529.823120   C1019.894958,526.722595 1010.358276,522.679138 1001.392029,517.862427   C989.873901,511.674835 983.924377,501.423431 981.816711,488.278748   C978.979431,470.584045 984.140869,455.662354 996.330200,443.436035   C1007.808899,431.922516 1022.358337,425.657562 1038.590454,424.089447   C1047.288086,423.249237 1056.135864,422.519165 1064.809204,423.166748   C1083.795166,424.584259 1101.731812,429.692749 1116.245972,442.859375   C1125.192627,450.975342 1122.850952,464.455444 1113.625610,470.594238   C1107.930542,474.383850 1101.479980,471.951935 1096.188721,468.537201   C1086.652466,462.383209 1076.847046,456.523041 1065.481201,455.415039   C1054.015747,454.297363 1042.304932,453.614014 1031.529053,459.850006   C1023.513611,464.488586 1018.134094,475.637207 1022.946838,483.790222   C1026.324707,489.512512 1031.919922,492.184052 1037.849487,493.947266   C1054.142334,498.792206 1070.557983,503.223846 1086.861206,508.034882   C1091.135986,509.296356 1095.178833,511.343414 1099.646484,513.241577  z" />
+      <path d=" M1207.028564,575.000000   C1228.687134,575.000000 1249.851196,575.065369 1271.014771,574.975098   C1283.322632,574.922546 1286.622314,582.621338 1286.805054,591.138123   C1286.998047,600.142883 1281.368652,606.059875 1270.853027,606.031494   C1238.690796,605.944824 1206.526733,605.824219 1174.367065,606.124634   C1166.408813,606.199036 1158.529297,600.774292 1157.983398,591.876709   C1157.733276,587.799377 1157.049805,583.731812 1157.043335,579.659058   C1156.973999,535.331665 1156.811157,491.002777 1157.124390,446.677490   C1157.209351,434.657806 1159.461914,429.048584 1170.911377,425.627136   C1172.440918,425.170044 1174.114990,425.026917 1175.722534,425.025146   C1206.885132,424.991211 1238.047852,425.062500 1269.210083,424.971100   C1282.814209,424.931213 1289.352539,434.614502 1285.921265,447.155212   C1284.507324,452.322693 1276.868042,456.986145 1269.938599,456.992401   C1247.441528,457.012726 1224.944580,457.000000 1202.447510,457.000000   C1200.808350,457.000000 1199.169189,457.000000 1197.266235,457.000000   C1197.266235,470.662018 1197.266235,484.066437 1197.266235,498.000000   C1201.854614,498.000000 1206.462158,498.000000 1211.069702,498.000000   C1228.067505,498.000000 1245.065918,497.915833 1262.062378,498.071838   C1265.410400,498.102570 1268.899536,498.616333 1272.061523,499.683624   C1278.148315,501.738068 1282.825684,509.536041 1281.962158,515.295532   C1280.564697,524.616272 1275.069092,529.950745 1266.474731,529.977356   C1244.477905,530.045471 1222.480591,530.000000 1200.483521,530.000061   C1199.504395,530.000061 1198.525146,530.000061 1197.278809,530.000061   C1197.278809,544.969543 1197.278809,559.707397 1197.278809,575.000000   C1200.319336,575.000000 1203.426636,575.000000 1207.028564,575.000000  z" />
+      <path d=" M980.000000,335.991394   C980.000000,292.665833 979.962097,249.838348 980.067078,207.011215   C980.076111,203.323029 980.521790,199.470917 981.657715,195.986099   C983.788696,189.448837 990.381653,185.034286 997.388062,185.022888   C1029.383545,184.970825 1061.379028,184.978668 1093.374634,185.016418   C1100.534058,185.024872 1107.084961,189.002304 1108.770630,194.462433   C1111.136475,202.126266 1110.090454,209.143982 1103.962158,213.282349   C1100.826050,215.400040 1096.554443,216.786255 1092.772095,216.840530   C1070.279907,217.163284 1047.780640,217.000000 1025.283813,217.000000   C1023.659424,217.000000 1022.035034,217.000000 1020.208374,217.000000   C1020.208374,230.780228 1020.208374,244.184021 1020.208374,258.000000   C1037.995117,258.000000 1055.611938,257.913055 1073.227173,258.045868   C1079.632446,258.094177 1086.059692,258.411560 1092.430176,259.063507   C1100.659180,259.905670 1105.432007,268.194153 1104.535645,276.283539   C1103.377075,286.739014 1095.406372,290.115631 1086.812622,290.050262   C1066.316650,289.894379 1045.818726,290.000000 1025.321533,290.000000   C1023.689148,290.000000 1022.056763,290.000000 1020.221680,290.000000   C1020.221680,305.086304 1020.221680,319.823334 1020.221680,335.000000   C1021.963135,335.000000 1023.579529,335.000000 1025.195923,335.000000   C1048.192749,335.000000 1071.189575,334.995300 1094.186401,335.002228   C1103.937866,335.005157 1110.086792,340.908081 1109.924072,350.803986   C1109.774658,359.882965 1104.599976,366.025970 1094.182129,366.013855   C1061.686646,365.976074 1029.191162,366.003845 996.695618,365.997498   C987.862122,365.995758 982.095886,361.049255 981.050598,352.199738   C980.435791,346.994263 980.333069,341.728302 980.000000,335.991394  z" />
+      <path d=" M1840.744141,543.334595   C1847.004395,556.933228 1852.989746,570.236084 1859.094116,583.484131   C1864.009155,594.151306 1856.761230,603.968811 1848.030640,606.352966   C1839.015991,608.814758 1830.241211,607.717285 1825.042725,596.890930   C1820.333740,587.083801 1816.052246,577.071472 1811.514893,567.000061   C1783.696655,567.000061 1755.882202,566.957153 1728.069092,567.129578   C1726.708374,567.138062 1724.765991,568.740295 1724.101318,570.094543   C1719.717163,579.027222 1715.380859,587.997986 1711.468506,597.144348   C1706.635254,608.444031 1695.131104,609.629089 1686.754883,605.595398   C1676.533081,600.672852 1672.601074,592.771423 1678.100342,580.816406   C1685.954712,563.741150 1694.415771,546.945984 1702.510010,529.979797   C1712.743652,508.529572 1722.871826,487.029114 1733.039551,465.547455   C1737.363403,456.412689 1741.773682,447.317017 1745.956177,438.117859   C1749.581787,430.143768 1755.225098,425.021332 1764.179443,423.990509   C1775.617676,422.673828 1784.810181,426.561005 1789.884277,436.960602   C1801.281860,460.320435 1812.043091,483.989929 1823.181396,507.477325   C1828.828857,519.386047 1834.732788,531.173157 1840.744141,543.334595  M1757.471924,495.020447   C1751.630371,508.162598 1745.788940,521.304810 1739.796875,534.785767   C1758.985718,534.785767 1777.741943,534.785767 1796.990479,534.785767   C1787.541992,513.510010 1778.132446,492.321747 1768.512573,470.660126   C1764.850952,478.902039 1761.406616,486.654968 1757.471924,495.020447  z" />
+      <path d=" M906.103638,186.975677   C928.509033,190.969360 944.776001,204.973846 949.036255,226.912811   C951.494141,239.570068 951.437622,252.081055 946.951355,264.227875   C941.054260,280.194763 929.578430,290.211395 913.257935,295.083649   C902.621277,298.259033 891.781555,298.908417 880.836914,298.982086   C869.046875,299.061432 857.255798,299.000000 845.000000,299.000000   C845.000000,315.922974 845.109009,332.526276 844.903015,349.125702   C844.864319,352.243683 844.057007,355.569885 842.756592,358.412109   C839.352234,365.852905 830.456238,369.826996 820.563110,367.877625   C812.318237,366.252991 806.049377,361.611603 805.028198,352.285950   C804.529663,347.733521 804.051331,343.148499 804.043762,338.577454   C803.970703,294.602570 803.807190,250.626114 804.126526,206.653442   C804.213623,194.656296 806.435608,189.067688 817.919556,185.619263   C819.447876,185.160324 821.123962,185.027557 822.732056,185.025055   C846.218628,184.988586 869.706055,184.905838 893.191284,185.087982   C897.353333,185.120255 901.506348,186.303207 906.103638,186.975677  M876.497559,267.000000   C880.826599,266.999969 885.164429,267.174011 889.482483,266.956635   C895.085754,266.674530 900.044067,264.649628 904.465332,261.068573   C912.789978,254.325974 913.111755,244.943375 911.684082,235.892990   C910.168457,226.284546 904.024536,219.687302 893.843872,217.994965   C890.487915,217.437134 887.072388,216.979782 883.679626,216.947754   C870.848450,216.826630 858.015442,216.899048 845.294434,216.899048   C845.294434,233.344574 845.294434,250.072403 845.294434,267.000000   C855.558105,267.000000 865.528259,267.000000 876.497559,267.000000  z" />
+      <path d=" M1603.005249,579.927856   C1600.586426,585.066711 1598.131470,589.781921 1595.835449,594.573242   C1591.112427,604.429321 1582.863403,608.415588 1572.480347,607.901672   C1562.888428,607.426880 1555.912842,602.792908 1551.875610,593.592773   C1546.910767,582.279297 1541.300049,571.250183 1536.012085,560.077576   C1523.304688,533.228088 1510.618774,506.368439 1497.932129,479.509216   C1493.269287,469.637390 1488.609009,459.764313 1483.983643,449.874878   C1479.739746,440.800903 1482.411499,431.929474 1490.934814,427.017975   C1501.862183,420.721191 1515.374268,422.910004 1521.125854,435.917755   C1533.707397,464.372162 1546.995728,492.513977 1559.976318,520.792236   C1564.847290,531.403564 1569.665527,542.039062 1574.766235,553.230591   C1583.648926,533.458008 1592.309570,513.988831 1601.136108,494.595184   C1607.717163,480.135529 1614.543457,465.787659 1621.196777,451.360535   C1623.611084,446.125183 1625.643188,440.706696 1628.203003,435.546967   C1631.510620,428.879395 1634.915527,422.765656 1649.557739,423.941345   C1655.568359,424.423981 1664.647095,430.276337 1665.561157,436.345215   C1666.253052,440.939331 1665.714355,446.351776 1663.860107,450.583801   C1656.901489,466.467010 1649.173828,482.013092 1641.743164,497.689667   C1629.432373,523.661987 1617.127563,549.637146 1604.824585,575.613098   C1604.209473,576.911865 1603.634766,578.229736 1603.005249,579.927856  z" />
+    </g>
+  </svg>
+);
