+++ frontend/src/components/Button.tsx
@@ -1,6 +1,5 @@
-import { Link } from '@tanstack/react-location';
 import clsx from 'clsx';
-import { AnchorHTMLAttributes, ButtonHTMLAttributes, DetailedHTMLProps, useRef } from 'react';
+import { ComponentPropsWithoutRef, ElementType, ReactNode, useRef } from 'react';
 import { AriaButtonProps, useButton, useFocusVisible } from 'react-aria';
 
 const styles = (isPressed = false) =>
@@ -10,32 +9,28 @@ const styles = (isPressed = false) =>
     isPressed && 'brightness-90',
   );
 
-type ButtonProps = { unstyled?: boolean } & (
-  | ({ href: string } & DetailedHTMLProps<AnchorHTMLAttributes<HTMLAnchorElement>, HTMLAnchorElement>)
-  | (Omit<AriaButtonProps<'button'>, 'href'> &
-      DetailedHTMLProps<ButtonHTMLAttributes<HTMLButtonElement>, HTMLButtonElement>)
-);
+type ButtonProps<C extends ElementType> =
+  | { as?: C; children: ReactNode; unstyled?: boolean } & AriaButtonProps<C> & ComponentPropsWithoutRef<C>;
 
-function Button({ unstyled, ...props }: ButtonProps) {
+export const Button = <C extends ElementType>({
+  as,
+  className,
+  children,
+  onPress,
+  unstyled,
+  ...props
+}: ButtonProps<C>) => {
+  const Component = as ?? 'button';
+  const ref = useRef(null);
+  const { buttonProps, isPressed } = useButton({ onPress, ...props }, ref);
   const { isFocusVisible } = useFocusVisible();
-  if ('href' in props) {
-    return <Link className={clsx(!unstyled && styles())} role="button" to={props.href} {...props} />;
-  }
-  const ref = useRef<HTMLButtonElement>(null);
-  const { buttonProps, isPressed } = useButton(props, ref);
   return (
-    <button
-      className={clsx(
-        !unstyled && styles(isPressed),
-        'outline-none',
-        isFocusVisible && 'focus:ring-4',
-        props.className,
-      )}
+    <Component
+      className={clsx(!unstyled && styles(isPressed), 'outline-none', isFocusVisible && 'focus:ring-4', className)}
+      {...props}
       {...buttonProps}
     >
-      {props.children}
-    </button>
+      {children}
+    </Component>
   );
-}
-
-export default Button;
+};
