+++ frontend/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Button';
+export * from './TextInput';
+export * from './Title';
